Stop Polling Game List When Game Starts
The GameBoard activity was getting launched each time the poller
updates the games and the current game is in progress, so the poller
needs to stop pulling games when the game actually starts, which makes
sense because in game play who cares about other games. When the game
ends in progress will get set to 0 and the poller will start polling
the gameList again.