Refactorization
I refactored the whole client side. Now the Fragments and Activities
use presenters to simplify their implementation. Also, when the DTO
comes back from the server, the ClientCommunicator now create command
on the client side and calls their .execute() function (This removed
all the passbackDTO functions from the MethodsFacade). Also, there is
now a ServerProxy which implements the same interface that the
serverFacade on the server side implements. This also cut a lot of code
out of the Methods facade so it should be a lot cleaner now :)